FIXUP: implement audio/video/spu delay
Each ESses will set the delay to their specific clocks.
Setting a delay on the master clock has an effect on all others slaves clocks.
Setting a delay on a slave clock has only an effect for the current clock.

For the audio master case (most common case), the aout need to wait (play
silence) when adding a positive delay to the master clock.