pulse: calculate the start time from the last date played
If I understand correctly, stream_latency_cb() is needed to re-calculate the
new derrered time (or start late in the worst case scenario) when the audio
device latency changes.

This callback called stream_start() with the first played system date. The
problem is that the delta calculated from stream_start was always late. Using
the last system date seems more logical. Indeed, the delta need to be
calculated from values using the same reference point (and vlc_pa_get_latency()
is calculated from the amount of buffers played).

This issue can be reproduced when the first played date is more than 100ms in
the future.